refactored api endpoints and cleanup
