+++ agenta-web/src/components/Playground/Playground.tsx
@@ -2,7 +2,8 @@ import React, {useState, useEffect, useRef} from "react"
 import {Button, Tabs, message} from "antd"
 import ViewNavigation from "./ViewNavigation"
 import NewVariantModal from "./NewVariantModal"
-import {fetchEnvironments, fetchVariants, saveNewVariant} from "@/lib/services/api"
+import {fetchVariants, saveNewVariant} from "@/lib/services/api"
+import {fetchEnvironments} from "@/services/deployment"
 import {Variant, PlaygroundTabsItem, Environment} from "@/lib/Types"
 import {AppstoreOutlined, SyncOutlined} from "@ant-design/icons"
 import {useRouter} from "next/router"
