+++ agenta-web/src/components/Playground/Views/TestView.tsx
@@ -1,7 +1,7 @@
 import React, {useContext, useEffect, useRef, useState} from "react"
 import {Button, Input, Card, Row, Col, Space, Form, Modal} from "antd"
 import {CaretRightOutlined, CloseCircleOutlined, PlusOutlined} from "@ant-design/icons"
-import {callVariant, promptRevision} from "@/lib/services/api"
+import {callVariant} from "@/lib/services/api"
 import {ChatMessage, ChatRole, GenericObject, Parameter, Variant} from "@/lib/Types"
 import {batchExecute, randString, removeKeys} from "@/lib/helpers/utils"
 import LoadTestsModal from "../LoadTestsModal"
@@ -23,6 +23,9 @@ import relativeTime from "dayjs/plugin/relativeTime"
 import duration from "dayjs/plugin/duration"
 import {useQueryParam} from "@/hooks/useQuery"
 import {formatLatency} from "@/lib/helpers/formatters"
+import {dynamicService} from "@/lib/helpers/dynamic"
+
+const promptRevision: any = dynamicService("api_ee")
 
 dayjs.extend(relativeTime)
 dayjs.extend(duration)
@@ -338,7 +341,7 @@ const App: React.FC<TestViewProps> = ({
             usage: {completion_tokens: number; prompt_tokens: number; total_tokens: number} | null
         }>
     >(testList.map(() => ({cost: null, latency: null, usage: null})))
-    const [revisionNum, setRevisionNum] = useQueryParam("revision")
+    const [revisionNum] = useQueryParam("revision")
 
     useEffect(() => {
         if (!revisionNum) return
