+++ agenta-web/src/components/TestSetTable/TestsetTable.tsx
@@ -4,7 +4,8 @@ import {createUseStyles} from "react-jss"
 import {Button, Input, Tooltip, Typography, message} from "antd"
 import TestsetMusHaveNameModal from "./InsertTestsetNameModal"
 import {DeleteOutlined, EditOutlined, PlusOutlined} from "@ant-design/icons"
-import {createNewTestset, fetchVariants, loadTestset, updateTestset} from "@/lib/services/api"
+import {fetchVariants} from "@/lib/services/api"
+import {createNewTestset, loadTestset, updateTestset} from "@/services/testsets"
 import {useRouter} from "next/router"
 import {useAppTheme} from "../Layout/ThemeContextProvider"
 import useBlockNavigation from "@/hooks/useBlockNavigation"
