+++ agenta-web/src/components/pages/evaluations/evaluationResults/NewEvaluationModal.tsx
@@ -2,8 +2,9 @@ import {useAppId} from "@/hooks/useAppId"
 import {JSSTheme, Variant, LLMRunRateLimit, testset} from "@/lib/Types"
 import {evaluatorConfigsAtom, evaluatorsAtom} from "@/lib/atoms/evaluation"
 import {apiKeyObject, redirectIfNoLLMKeys} from "@/lib/helpers/utils"
-import {fetchTestsets, fetchVariants} from "@/lib/services/api"
+import {fetchVariants} from "@/lib/services/api"
 import {CreateEvaluationData, createEvalutaiton} from "@/services/evaluations"
+import {fetchTestsets} from "@/services/testsets"
 import {PlusOutlined, QuestionCircleOutlined} from "@ant-design/icons"
 import {
     Divider,
