+++ agenta-web/src/pages/apps/[app_id]/annotations/single_model_test/[evaluation_id]/index.tsx
@@ -1,5 +1,6 @@
 import {Evaluation, EvaluationScenario, GenericObject} from "@/lib/Types"
-import {loadEvaluation, loadEvaluationsScenarios, loadTestset} from "@/lib/services/api"
+import {loadEvaluation, loadEvaluationsScenarios} from "@/services/human-evaluations"
+import {loadTestset} from "@/services/testsets"
 import {useRouter} from "next/router"
 import {useEffect, useState} from "react"
 import {fetchVariants} from "@/lib/services/api"
