+++ agenta-web/src/pages/apps/[app_id]/testsets/index.tsx
@@ -5,7 +5,7 @@ import {ColumnsType} from "antd/es/table"
 import {useState} from "react"
 import {formatDate} from "@/lib/helpers/dateTimeHelper"
 import {DeleteOutlined} from "@ant-design/icons"
-import {deleteTestsets, useLoadTestsetsList} from "@/lib/services/api"
+import {deleteTestsets, useLoadTestsetsList} from "@/services/testsets"
 import {createUseStyles} from "react-jss"
 import {testset} from "@/lib/Types"
 import {isDemo} from "@/lib/helpers/utils"
